HealthMLBase.jl

This is the documentation for HealthMLBase.

HealthMLBase.fit!

HealthMLBase.inverse_transform!

HealthMLBase.predict!

HealthMLBase.predict_log_proba

HealthMLBase.predict_log_proba!

HealthMLBase.predict_proba

HealthMLBase.predict_proba!

HealthMLBase.transform!

MLJModelInterface._model_cleaner

MLJModelInterface._model_constructor

MLJModelInterface._process_model_def

MLJModelInterface._unpack!

MLJModelInterface.classes

MLJModelInterface.decoder

MLJModelInterface.docstring_ext

MLJModelInterface.evaluate

MLJModelInterface.fit

MLJModelInterface.int

MLJModelInterface.inverse_transform

MLJModelInterface.is_same_except

MLJModelInterface.isrepresented

MLJModelInterface.matrix

MLJModelInterface.metadata_model

MLJModelInterface.metadata_pkg

MLJModelInterface.nrows

MLJModelInterface.predict

MLJModelInterface.predict_mean

MLJModelInterface.predict_median

MLJModelInterface.predict_mode

MLJModelInterface.schema

MLJModelInterface.select

MLJModelInterface.selectcols

MLJModelInterface.selectrows

MLJModelInterface.table

MLJModelInterface.transform

MLJModelInterface.update

MLJModelInterface.update_data

MLJModelInterface.@mlj_model

HealthMLBase.fit! — Function

fit!

HealthMLBase.inverse_transform! — Function

inverse_transform!

HealthMLBase.predict! — Function

predict!

HealthMLBase.predict_log_proba — Function

predict_log_proba

HealthMLBase.predict_log_proba! — Function

predict_log_proba!

HealthMLBase.predict_proba — Function

predict_proba

HealthMLBase.predict_proba! — Function

predict_proba!

HealthMLBase.transform! — Function

transform!

MLJModelInterface.classes — Method

classes(x)

All the categorical elements with the same pool as x (including x), returned as a list, with an ordering consistent with the pool (requires MLJBase to be loaded). Here x has CategoricalValue or CategoricalString type, and classes(x) is a vector of the same eltype. Note that x in classes(x) is always true.

Not to be confused with levels(x.pool). See the example below.

julia>  v = categorical([:c, :b, :c, :a])
4-element CategoricalArrays.CategoricalArray{Symbol,1,UInt32}:
 :c
 :b
 :c
 :a

julia> levels(v)
3-element Array{Symbol,1}:
 :a
 :b
 :c

julia> x = v[4]
CategoricalArrays.CategoricalValue{Symbol,UInt32} :a

julia> classes(x)
3-element CategoricalArrays.CategoricalArray{Symbol,1,UInt32}:
 :a
 :b
 :c

julia> levels(x.pool)
3-element Array{Symbol,1}:
 :a
 :b
 :c

MLJModelInterface.decoder — Method

d = decoder(x)

A callable object for decoding the integer representation of a CategoricalString or CategoricalValue sharing the same pool as x (requires MLJBase to be loaded). (Here x is of one of these two types.) Specifically, one has d(int(y)) == y for all y in classes(x). One can also call d on integer arrays, in which case d is broadcast over all elements.

julia> v = categorical([:c, :b, :c, :a])
julia> int(v)
4-element Array{UInt32,1}:
 0x00000003
 0x00000002
 0x00000003
 0x00000001
julia> d = decoder(v[3])
julia> d(int(v)) == v
true

Warning: It is not true that int(d(u)) == u always holds.

See also: int, classes.

MLJModelInterface.evaluate — Function

some meta-models may choose to implement the evaluate operations

MLJModelInterface.fit — Function

every model interface must implement a fit method of the form fit(model, verb::Integer, training_args...) -> fitresult, cache, report

MLJModelInterface.int — Method

int(x; type=nothing)

The positional integer of the CategoricalString or CategoricalValue x, in the ordering defined by the pool of x. The type of int(x) is the reference type of x (requires MLJBase to be loaded).

Not to be confused with x.ref, which is unchanged by reordering of the pool of x, but has the same type.

int(X::CategoricalArray)
int(W::Array{<:CategoricalString})
int(W::Array{<:CategoricalValue})

Broadcasted versions of int.

julia> v = categorical([:c, :b, :c, :a])
julia> levels(v)
3-element Array{Symbol,1}:
 :a
 :b
 :c
julia> int(v)
4-element Array{UInt32,1}:
 0x00000003
 0x00000002
 0x00000003
 0x00000001

See also: decoder.

MLJModelInterface.inverse_transform — Function

unsupervised methods may implement the inverse_transform operation

MLJModelInterface.is_same_except — Method

is_same_except(m1::MLJType, m2::MLJType, exceptions::Symbol...)

Returns true only the following conditions all hold:

m1 and m2 have the same type.

m1 and m2 have the same undefined fields.

Corresponding fields agree, or are listed as exceptions, or have AbstractRNG as values (one or both)

Here "agree" is in the sense of "==", unless the objects are themselves of MLJType, in which case agreement is in the sense of is_same_except with no exceptions allowed.

Note that Base.== is overloaded such that m1 == m2 if and only if is_same_except(m1, m2).

MLJModelInterface.isrepresented — Method

isrepresented(object::MLJType, objects)

Test if object has a representative in the iterable objects. This is a weaker requirement than object in objects.

Here we say m1 respresents m2 if is_same_except(m1, m2) is true.

MLJModelInterface.matrix — Method

matrix(X; transpose=false)

If X <: AbstractMatrix, return X or permutedims(X) if transpose=true. If X is a Tables.jl compatible table source, convert X into a Matrix (requires MLJBase to be loaded).

MLJModelInterface.metadata_model — Method

metadata_model(`T`; args...)

Helper function to write the metadata for a model T.

Keywords

input=Unknown : allowed scientific type of the input data

target=Unknown: allowed sc. type of the target (supervised)

output=Unknown: allowed sc. type of the transformed data (unsupervised)

weights=false : whether the model supports sample weights

descr="" : short description of the model

path="" : where the model is (usually PackageName.ModelName)

Example

metadata_model(KNNRegressor,
    input=MLJModelInterface.Table(MLJModelInterface.Continuous),
    target=AbstractVector{MLJModelInterface.Continuous},
    weights=true,
    descr="K-Nearest Neighbors classifier: ...",
    path="NearestNeighbors.KNNRegressor")

MLJModelInterface.metadata_pkg — Method

metadata_pkg(T; args...)

Helper function to write the metadata for a package providing model T. Use it with broadcasting to define the metadata of the package providing a series of models.

Keywords

name="unknown" : package name

uuid="unknown" : package uuid

url="unknown" : package url

julia=missing : whether the package is pure julia

license="unknown": package license

is_wrapper=false : whether the package is a wrapper

Example

metadata_pkg.((KNNRegressor, KNNClassifier),
    name="NearestNeighbors",
    uuid="b8a86587-4115-5ab1-83bc-aa920d37bbce",
    url="https://github.com/KristofferC/NearestNeighbors.jl",
    julia=true,
    license="MIT",
    is_wrapper=false)

MLJModelInterface.nrows — Method

nrows(X)

Return the number of rows for a table, abstract vector or matrix X (requires MLJBase to be loaded).

MLJModelInterface.predict — Function

supervised methods must implement the predict operation

MLJModelInterface.predict_mean — Function

probabilistic supervised models may overload predict_mean

MLJModelInterface.predict_median — Function

probabilistic supervised models may overload predict_median

MLJModelInterface.predict_mode — Function

probabilistic supervised models may overload predict_mode

MLJModelInterface.select — Method

select(X, r, c)

Select element(s) of a table or matrix at row(s) r and column(s) c. An object of the sink type of X (or a matrix) is returned unless c is a single integer or symbol. In that case a vector is returned, unless r is a single integer, in which case a single element is returned.

See also: selectrows, selectcols.

MLJModelInterface.selectcols — Method

selectcols(X, c)

Select single or multiple columns from a matrix or table X (requires MLJBase to be loaded). If c is an abstract vector of integers or symbols, then the object returned is a table of the preferred sink type of typeof(X). If c is a single integer or column, then an AbstractVector is returned.

MLJModelInterface.selectrows — Method

selectrows(X, r)

Select single or multiple rows from a table, abstract vector or matrix X (requires MLJBase to be loaded). If X is tabular, the object returned is a table of the preferred sink type of typeof(X), even if only a single row is selected.

MLJModelInterface.table — Method

table(columntable; prototype=nothing)

Convert a named tuple of vectors or tuples columntable, into a table of the "preferred sink type" of prototype (requires MLJBase to be loaded). This is often the type of prototype itself, when prototype is a sink; see the Tables.jl documentation. If prototype is not specified, then a named tuple of vectors is returned.

table(A::AbstractMatrix; names=nothing, prototype=nothing)

Wrap an abstract matrix A as a Tables.jl compatible table with the specified column names (a tuple of symbols). If names are not specified, names=(:x1, :x2, ..., :xn) is used, where n=size(A, 2) (requires MLJBase to be loaded).

If a prototype is specified, then the matrix is materialized as a table of the preferred sink type of prototype, rather than wrapped. Note that if prototype is not specified, then matrix(table(A)) is essentially a no-op.

MLJModelInterface.transform — Function

unsupervised methods must implement the transform operation

MLJModelInterface.update — Method

each model interface may overload the update refitting method

MLJModelInterface.update_data — Function

each model interface may overload the update_data refitting method for online learning

MLJModelInterface.@mlj_model — Macro

@mlj_model

Macro to help define MLJ models with constraints on the default parameters.

MLJModelInterface._model_cleaner — Method

_model_cleaner(modelname, defaults, constraints)

Build the expression of the cleaner associated with the constraints specified in a model def.

MLJModelInterface._model_constructor — Method

_model_constructor(modelname, params, defaults)

Build the expression of the keyword constructor associated with a model definition. When the constructor is called, the clean! function is called as well to check that parameter assignments are valid.

MLJModelInterface._process_model_def — Method

_process_model_def(ex)

Take an expression defining a model (mutable struct Model ...) and unpack key elements for further processing:

Model name (modelname)

Names of parameters (params)

Default values (defaults)

Constraints (constraints)

MLJModelInterface._unpack! — Method

_unpack!(ex, rep)

Internal function to allow to read a constraint given after a default value for a parameter and transform it in an executable condition (which is returned to be executed later). For instance if we have

alpha::Int = 0.5::(arg > 0.0)

Then it would transform the (arg > 0.0) in (alpha > 0.0) which is executable.

MLJModelInterface.docstring_ext — Method

docstring_ext

Internal function to help generate the docstring for a package. See metadata_model.

MLJModelInterface.schema — Method

schema(X)

If X is tabular, inspect the column types and scitypes, otherwise return nothing.